REFACTOR : 검색어 관련 기능 Service & Controller 생성
