+++ src/test/java/com/Bridge/bridge/service/ProjectServiceTest.java
@@ -1224,81 +1224,6 @@ class ProjectServiceTest {
         Assertions.assertThat(bookmarkResponseDto.getScrap()).isEqualTo("스크랩이 해제되었습니다.");
     }
 
-    @DisplayName("최근 검색어 조회")
-    @Test
-    void resentSearchWord() {
-        // given
-        User user = new User("user1", "user1@gmail.com", Platform.APPLE, "Test");
-        user = userRepository.save(user);
-
-        SearchWord newSearch1 = SearchWord.builder()
-                .content("검색어1")
-                .user(user)
-                .history(LocalDateTime.now())
-                .build();
-        SearchWord newSearch2 = SearchWord.builder()
-                .content("검색어2")
-                .user(user)
-                .history(LocalDateTime.now())
-                .build();
-        SearchWord newSearch3 = SearchWord.builder()
-                .content("검색어3")
-                .user(user)
-                .history(LocalDateTime.now())
-                .build();
-
-        searchWordRepository.save(newSearch1);
-        searchWordRepository.save(newSearch2);
-        searchWordRepository.save(newSearch3);
-
-        user.getSearchWords().add(newSearch1);
-        user.getSearchWords().add(newSearch2);
-        user.getSearchWords().add(newSearch3);
-
-        // when
-        List<SearchWordResponseDto> searchWordResponseDto = projectService.resentSearchWord(user.getId());
-        Assertions.assertThat(searchWordResponseDto.get(0).getSearchWord()).isEqualTo("검색어1");
-        Assertions.assertThat(searchWordResponseDto.get(1).getSearchWord()).isEqualTo("검색어2");
-        Assertions.assertThat(searchWordResponseDto.get(2).getSearchWord()).isEqualTo("검색어3");
-    }
-
-    @DisplayName("최근 검색어 삭제")
-    @Test
-    void deleteSearchWord() {
-        // given
-        User user = new User("user1", "user1@gmail.com", Platform.APPLE, "Test");
-        user = userRepository.save(user);
-
-        SearchWord newSearch1 = SearchWord.builder()
-                .content("검색어1")
-                .user(user)
-                .history(LocalDateTime.now())
-                .build();
-        SearchWord newSearch2 = SearchWord.builder()
-                .content("검색어2")
-                .user(user)
-                .history(LocalDateTime.now())
-                .build();
-        SearchWord newSearch3 = SearchWord.builder()
-                .content("검색어3")
-                .user(user)
-                .history(LocalDateTime.now())
-                .build();
-
-        searchWordRepository.save(newSearch1);
-        searchWordRepository.save(newSearch2);
-        searchWordRepository.save(newSearch3);
-
-        user.getSearchWords().add(newSearch1);
-        user.getSearchWords().add(newSearch2);
-        user.getSearchWords().add(newSearch3);
-
-        // when
-        List<SearchWordResponseDto> searchWordResponseDto = projectService.deleteSearchWord(user.getId(), newSearch1.getId());
-        Assertions.assertThat(searchWordResponseDto.get(0).getSearchWord()).isEqualTo("검색어2");
-        Assertions.assertThat(searchWordResponseDto.get(1).getSearchWord()).isEqualTo("검색어3");
-    }
-
     @Test
     @DisplayName("프로젝트 지원 취소하기")
     @Transactional
